split interaction module into zone and manager
