+++ pets-lib/src/world/interaction/mod.rs
@@ -1,33 +1,6 @@
 //!
-//! Area that allows the player to start
-//! an interaction when within range
+//! Stuff for interactions in the overworld
 //!
 
-use godot::prelude::*;
-
-use godot::engine::{Area2D, Area2DVirtual};
-
-#[derive(GodotClass)]
-#[class(base=Area2D)]
-struct InteractionZone {
-    #[base]
-    node: Base<Area2D>,
-
-    #[export]
-    name: GodotString,
-}
-
-#[godot_api]
-impl InteractionZone {}
-
-#[godot_api]
-impl Area2DVirtual for InteractionZone {
-    fn init(node: Base<Area2D>) -> Self {
-        Self {
-            node,
-            name: "".into(),
-        }
-    }
-
-    // fn ready(&mut self) {}
-}
+mod manager;
+mod zone;
